fix: Modify the update() logic and remove unnecessary methods.
Now the updateHooK() logic is basically related to the user action
check, first of all, and them, check the vehicles states (like the
vehicle cmds are anync and the user can get the control authority
anytime, this is mandatory) before to setting the cmd to the vehicle.
Further, the name of the methods have been modified
(land, takeoff, mission, preland).